kill program with chris 1  19/11/2013

+++ sketchbook/jamie_motor_driver_nano/motor_driver_kill_02/motor_driver_kill_02.ino
@@ -77,14 +77,14 @@ void motors_off(void)		//function sets all the motor speeds to 0.
 
 void lcdLine1CallBack( const std_msgs::String& msg)	//sets up the first line of the LCD?
  {
-  //lcd.setCursor(0, 0); 					// set the cursor to column 0, line 0.
-  //lcd.print(msg.data);
+  lcd.setCursor(0, 0); 					// set the cursor to column 0, line 0.
+  lcd.print(msg.data);
  }
 
 void lcdLine2CallBack( const std_msgs::String& msg)	//sets up the second line of the LCD?
  {
- // lcd.setCursor(0, 1);					// set the cursor to column 0, line 1.
- // lcd.print(msg.data);
+  lcd.setCursor(0, 1);					// set the cursor to column 0, line 1.
+  lcd.print(msg.data);
  }
 
 int averageAnalog(int pin)				//function is used for calculating battery voltage.
@@ -95,25 +95,27 @@ int averageAnalog(int pin)				//function is used for calculating battery voltage
  }
 
 
-void guiEmergencyCallBack( const std_msgs::Bool& eFlag) //function checks for a change of the emergency kill flag.
+/*void guiEmergencyCallBack( const std_msgs::Bool& eFlag) //function checks for a change of the emergency kill flag.
  {
- /*  if (eFlag.data) //if flag is set to 1
+   if (eFlag.data) //if flag is set to 1
      {
        emergencyKill = true; //emergency kill triggered.
      }
    if(!eFlag.data) //if flag is set to 0
      {
        emergencyKill = false; //emergeny kill not triggered.
-     }*/
+     }
  }
- 
+ */
  ros::Publisher m("batteryStatusMotor", &batteryStatusMotor);				//publishes the motor battery status.
  ros::Publisher s("batteryStatusSystem", &batteryStatusSystem);				//publiches the system battery status.
  
  ros::Subscriber<custom_msg::MotorConfig> sub("motor_config", &motorConfigCallBack ); 	//subscribes to the motor config input.
+ 
  ros::Subscriber<std_msgs::String> sub1("lcd_line_1", &lcdLine1CallBack );		//subscribes to recieve a string for the LCD.
  ros::Subscriber<std_msgs::String> sub2("lcd_line_2", &lcdLine2CallBack );
- ros::Subscriber<std_msgs::Bool> emergency("emergency", &guiEmergencyCallBack );//subscribes to recieve a boolian value for the emergency kill.
+ 
+ //ros::Subscriber<std_msgs::Bool> emergency("emergency", &guiEmergencyCallBack );//subscribes to recieve a boolian value for the emergency kill.
 
 
 void setup() 
@@ -128,7 +130,7 @@ void setup()
    nh.subscribe(sub); //sets the node handles for the subscribers
    nh.subscribe(sub1);
    nh.subscribe(sub2);
-   nh.subscribe(emergency);
+  // nh.subscribe(emergency);
   
    nh.advertise(m);  //sets the node handles for the publishers.
    nh.advertise(s);
